+++ HeroesApi/Models/Contracts/IDataRepository.cs
@@ -4,7 +4,7 @@ namespace Models.Contracts;
 public interface IDataRepository
 {
     Task<IEnumerable<Hero>> GetAllAsync();
-    Task<Hero?> GetAsync(int id);
+    Task<Hero?> GetAsync(long id);
     Task AddAsync(Hero entity);
     Task UpdateAsync(Hero entityToUpdate, Hero entity);
     Task DeleteAsync(Hero entity);
